move SSG logic out of webpack
